plot update
